# ns-topup/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NS Games Topup API",
        "description": "API untuk platform topup game, pulsa, dan e-wallet dengan integrasi Digiflazz dan Midtrans",
        "contact": {
            "name": "NS Games Support",
            "email": "admin@nsgames.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Laravel Development Server"
        },
        {
            "url": "http://localhost/NS-topupgames/ns-topup/public",
            "description": "Laragon/XAMPP Server"
        }
    ],
    "paths": {
...
        "/api/admin/products": {
            "get": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "Get all products for admin",
                "operationId": "9058305dfe0377e6cf7c5253ed664401",
                "responses": {
                    "200": {
                        "description": "List of all products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "1_1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Mobile Legends - 86 Diamonds"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Game"
                                                    },
                                                    "provider": {
                                                        "type": "string",
                                                        "example": "Digiflazz"
                                                    },
                                                    "buyPrice": {
                                                        "type": "number",
                                                        "example": 20000
                                                    },
                                                    "sellPrice": {
                                                        "type": "number",
                                                        "example": 22000
                                                    },
                                                    "profit": {
                                                        "type": "number",
                                                        "example": 2000
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "active"
                                                    },
                                                    "stock": {
                                                        "type": "string",
                                                        "example": "unlimited"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "Create new product",
                "operationId": "e78a4cd5a8e47d96daa7a5a7a13d3e30",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "game_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_name": {
                                        "type": "string",
                                        "example": "86 Diamonds"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 86
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 22000
                                    },
                                    "base_price": {
                                        "type": "number",
                                        "example": 20000
                                    },
                                    "sku_code": {
                                        "type": "string",
                                        "example": "ML_86"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/{id}": {
            "get": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "Get product details",
                "operationId": "e645351e64571486935b1fbb5743d6b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "Update product",
                "operationId": "78c34d6552e97335a6594ceaf054d60f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "example": "86 Diamonds"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 86
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 22000
                                    },
                                    "base_price": {
                                        "type": "number",
                                        "example": 20000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Products"
                ],
                "summary": "Delete product",
                "operationId": "14131d6e77dc1537765eaa35848d3941",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}